-added button hover to dash page, removed duplicate product page file, there are now 3 template files for dashboards that need to be modified
