+++ src/app/index.scss
@@ -389,7 +389,6 @@ body {
     display: flex;
     flex-wrap: wrap;
     justify-content: center;
-    
   }
 
   * > .activity {
@@ -432,11 +431,6 @@ body {
     margin: 20px;
     border: 1px solid #9700ff;
     border-radius: 50px;
-    }
-
-  .upload_btn {
-    margin-left: 100px;
-    padding-bottom: 50px;
     }
 
   .about {
@@ -496,21 +490,15 @@ body {
     color: black;
     border-radius: 50px;
     background-image: linear-gradient(#9700ff, aqua, blue, #9700ff);  
-  
-    
   }
 
-    .buttonaf {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 10px;
-    justify-content: space-evenly;
-    margin: 20px;
-    max-width: 250px;
+  .dash-btn:hover {
+    border: 1px solid #ffffff;
+    box-shadow: 0 0 5px 0 #ffffff;
   }
+
   /* =======================
-  THEME PICKER
+       THEME PICKER
   =======================*/
 
   .theme_area_header {
@@ -548,14 +536,14 @@ body {
     margin-bottom: 20px;
   }
 
-    .preview {
+  .preview {
     width: 100px;
     height: 100px;
     margin: 20px auto;
     box-shadow: 0px 0px 15px 7px #492951;
   }
 
-   .colorpickerheading{
+  .colorpickerheading{
     font-family: "cabin", sans-serif;
     font-size: 30px;
     color: #9700ff;
@@ -568,12 +556,6 @@ body {
     border: 10px solid transparent;
     border-radius: 10px;
   }
-
-  /** Text Shadow **/
-  .text-shadow {
-    text-shadow: 2px 2px 2px #000000;
-  }
-
 
   /* =======================
         COLOR SLIDERS - RGB
